% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcInfo.R
\name{calcInfo}
\alias{calcInfo}
\title{calcInfo.R}
\usage{
calcInfo(
  leftDmat,
  rightDmat,
  gSummed,
  gMat = NULL,
  leftTimes,
  theta1,
  theta2,
  deltaVec
)
}
\arguments{
\item{leftDmat}{n*(p+nknots+2) design matrix for left end of interval.}

\item{rightDmat}{n*(p+nknots+2) design matrix for right end of interval.}

\item{gSummed}{n*1 vector of summed genotype matrix.}

\item{gMat}{n*q genotype matrix (default is NULL).}

\item{leftTimes}{n*1 vector of left side of interval times.}

\item{theta1}{n*1 vector of fitted coefficient for cause 1.}

\item{theta2}{n*1 vector of fitted coefficient for cause 2.}

\item{deltaVec}{n*1 vector of the event cause (cause 1 or 2; or right-censored 0).}
}
\value{
A Fisher information matrix for model coefficients
}
\description{
Calculate Fisher information matrix.
}
\examples{
seed <- 2022
n <- 3000
q <- 20
alpha1 <- -0.058
alpha2 <- -0.035
beta1 <- 0.03
beta2 <- log(1 - exp(beta1 / alpha1)) * alpha2
outcomeDat <- genData(seed=seed, n=n, alpha1=alpha1, alpha2=alpha2, beta1=beta1, beta2=beta2)
xMat <- matrix(data=rnorm(n), nrow=n)
gMat <- matrix(data=rbinom(n=n*q, size=2, prob=0.3), ncol=q)
gSummed <- matrix(data=apply(gMat, 1, sum), ncol=1)
obsTimes <- 1:5
lt <- outcomeDat$leftTimes
rt <- outcomeDat$rightTimes
obsInd <- outcomeDat$deltaVecSimple
dmats <- makeICdmat(xMat = xMat, lt = lt, rt = rt, obsInd = obsInd, quant_r = NULL, nKnots = 1)
iMatPartial <- calcInfo(leftDmat = dmats$left_dmat, rightDmat = dmats$right_dmat, leftTimes = lt,
theta1 = runif(4, 0.01, 0.02), theta2 = runif(5, 0.01, 0.02), deltaVec = outcomeDat$deltaVec,
gSummed=gSummed, gMat = NULL)

}
