% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIntNew.R
\name{createIntNew}
\alias{createIntNew}
\title{createIntNew.R}
\usage{
createIntNew(obsTimes, eventTime)
}
\arguments{
\item{obsTimes}{A vector of all the times a subject is observed.}

\item{eventTime}{The exact event time for the subject.}
}
\value{
A 2*1 vector which is the interval of the event time.
}
\description{
Turn the actual observed outcome times and observation times into interval-censored
outcomes for each subject. Apply this with mapply over a data.frame of visit times, pass in the exact times.
}
\examples{
obsTimes <- 1:10
eventTime <- 4.4
createIntNew(obsTimes, eventTime)

}
