% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crSKAT.R
\name{crSKAT}
\alias{crSKAT}
\title{crSKAT.R}
\usage{
crSKAT(
  leftDmat,
  rightDmat,
  leftTimes,
  deltaVec,
  gMat,
  gSummed,
  null_beta,
  pvalue = TRUE
)
}
\arguments{
\item{leftDmat}{n*(p+nknots+2) design matrix for left end of interval.}

\item{rightDmat}{n*(p+nknots+2) design matrix for right end of interval.}

\item{leftTimes}{n*1 vector of left side of interval times.}

\item{deltaVec}{n*1 vector of the event cause (cause 1 or 2; or right-censored 0).}

\item{gMat}{n*q genotype matrix.}

\item{gSummed}{n*1 vector of summed genotype matrix.}

\item{null_beta}{(p+nknots+2)*1 vector of coefficients for null model.}

\item{pvalue}{If TRUE, then find the p-value}
}
\value{
A list with the elements:
\item{p_SKAT}{ICSKAT p-value}
\item{p_burden}{IC burden test p-value}
\item{complex}{Indicator of whether the SKAT variance matrix was positive definite}
\item{sig_mat}{The covariance matrix of the score equations for genetic effects when treated as fixed effects}
\item{skatQ}{SKAT test statistic}
\item{burdenQ}{Burden test statistic}
\item{Ugamma}{Score vector}
\item{lambdaQ}{Vector of eigenvalues of variance matrix}
\item{null_beta}{The fitted null parameters}
\item{err}{Will be 0 for no error, 22 if had to adjust parameters on CompQuadForm (totally normal), or 99 if NA in variance matrix. ICSKATwrapper will return 1 here if the null fit has an error}
\item{errMsg}{Explains error code, blank string if no error}
}
\description{
Calculate the test statistic and p-value for interval-censored competing risks SKAT.
}
\examples{
n <- 30000
q <- 500
alpha1 <- -0.1
alpha2 <- -0.3
beta1 <- 0.03
beta2 <- log(1 - exp(beta1 / alpha1)) * alpha2
outcomeDat <- genData(seed=NULL, n=n, alpha1=alpha1, alpha2=alpha2, beta1=beta1, beta2=beta2)
xMat <- matrix(data=rnorm(n), nrow=n)
gMat <- matrix(data=rbinom(n=n*q, size=2, prob=0.3), ncol=q)
gSummed <- matrix(data=apply(gMat, 1, sum), ncol=1)
obsTimes <- 3:7
lt <- outcomeDat$leftTimes
rt <- outcomeDat$rightTimes
obsInd <- outcomeDat$deltaVecSimple
deltaVec <- outcomeDat$deltaVec
dmats <- makeICdmat(xMat = xMat, lt = lt, rt = rt, obsInd = obsInd, quant_r = NULL, nKnots = 1)
leftDmat <- dmats$left_dmat
rightDmat <- dmats$right_dmat
nullFit <- crSKAT_fit_null(init_beta=rep(0, 9), leftDmat=leftDmat, rightDmat=rightDmat,
deltaVec=deltaVec, leftTimes=lt, gSummed=gSummed, method="Broyden")
out <- crSKAT(leftDmat=dmats$left_dmat, rightDmat=rightDmat, leftTimes=lt,
deltaVec=deltaVec, gMat=gMat, gSummed=gSummed, null_beta=nullFit$beta_fit, pvalue=TRUE)

}
