% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Simulated data using Gompertz distribution}
\format{
A list of 2 with 1000 observations
\describe{
\item{tempTime}{The time-to-event observed time.}
\item{tempType}{The type of causes to the outcome: primary outcome (1) or competing event (2).}
}
}
\usage{
exampleData
}
\description{
simulated time-to-event outcome and event type
}
\references{
\url{https://onlinelibrary.wiley.com/doi/10.1111/biom.12109}
}
\keyword{datasets}
