% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_GMat.R
\name{gen_GMat}
\alias{gen_GMat}
\title{gen_GMat.R}
\usage{
gen_GMat(n, q, rho)
}
\arguments{
\item{n}{the number of observations to generate (sample size).}

\item{q}{the number of SNPs.}

\item{rho}{the correlation for the binary correlation matrix.}
}
\value{
A n*q correlated matrix for simulated SNPs.
}
\description{
Generate the multivariate binary random genotype matrix for simulation studies.
}
\examples{
n <- 1000
q <- 20
rho <- 0.1
gen_GMat(n, q, rho)


}
