% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genData.R
\name{genData}
\alias{genData}
\title{genData.R}
\usage{
genData(
  seed = NULL,
  n,
  alpha1,
  alpha2,
  beta1,
  beta2,
  obsTimes = seq(from = 4, to = 28, by = 4),
  probMiss = 0.1,
  windowHalf = 1
)
}
\arguments{
\item{seed}{Random seed.}

\item{n}{The number of observations to generate (sample size).}

\item{alpha1}{First parameter for Gompertz model.}

\item{alpha2}{Second parameter for Gompertz model.}

\item{beta1}{Third parameter for Gompertz model.}

\item{beta2}{Fourth parameter for Gompertz model.}

\item{obsTimes}{Vector of the intended observation times.}

\item{probMiss}{The probability of missing any given visit (default is 0.1).}

\item{windowHalf}{The amount of time before or after the intended obsTimes that a visit might take place (default is 1).}
}
\value{
A list with the elements:
\item{tempTime}{n\emph{1 vector of exact event times.}
\item{tempType}{n}1 vector of exact event types}
\item{deltaVecSimple}{n\emph{1 vector of whether the event was observed after follow-up started (t>0).}
\item{deltaVec}{n}1 vector of the event cause (cause 1 or 2; or right-censored 0).}
\item{leftTimes}{n\emph{1 vector of left side of interval times.}
\item{rightTimes}{n}1 vector of right side of interval times.}
}
\description{
Generate competing risks interval-censored data under the proportional odds given a baseline hazard function and
some information about observation times.
}
\examples{
seed <- 2022
n <- 1000
alpha1 <- -0.058
alpha2 <- -0.035
beta1 <- 0.03
beta2 <- log(1 - exp(beta1 / alpha1)) * alpha2
outcomeDat <- genData(seed=seed, n=n, alpha1=alpha1, alpha2=alpha2, beta1=beta1, beta2=beta2,
obsTimes=seq(from=4, to=28, by=4), probMiss=0.1, windowHalf=1)

}
