% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genData_pwr.R
\name{genData_pwr}
\alias{genData_pwr}
\title{genData_pwr.R}
\usage{
genData_pwr(
  n,
  alpha1,
  alpha2,
  beta1,
  beta2,
  gMatCausal,
  effectSizes,
  seed = NULL
)
}
\arguments{
\item{n}{the number of observations to generate (sample size).}

\item{alpha1}{1st parameter for Gompertz model.}

\item{alpha2}{2nd parameter for Gompertz model.}

\item{beta1}{3rd parameter for Gompertz model.}

\item{beta2}{4th parameter for Gompertz model.}

\item{gMatCausal}{the genotype matrix of causal SNPs.}

\item{effectSizes}{the effect size of each causal SNP.}

\item{seed}{random seed.}
}
\value{
A list with the elements:
\item{tempTime}{A vector of exact event times.}
\item{tempType}{A vector of exact event types.}
}
\description{
Generate the data for power simulation studies.
}
\examples{
n <- 1000
alpha1 <- -0.058
alpha2 <- -0.035
beta1 <- 0.03
beta2 <- log(1 - exp(beta1 / alpha1)) * alpha2
gMatCausal <- gen_GMat(n, q=5, rho=0.1)
effectSizes <- seq(0.1, 0.5, length.out=5)
genData_pwr(n, alpha1, alpha2, beta1, beta2, gMatCausal, effectSizes, seed=NULL)


}
