% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeICdmat.R
\name{makeICdmat}
\alias{makeICdmat}
\title{makeICdmat.R}
\usage{
makeICdmat(xMat, lt, rt, obsInd, quant_r = NULL, nKnots = 1)
}
\arguments{
\item{xMat}{n*p matrix of non-genetic covariates.}

\item{lt}{n*1 vector with left end of intervals (min is 0).}

\item{rt}{n*1 vector with right end of intervals.}

\item{obsInd}{n*1 vector of whether the event was observed before last follow-up.}

\item{quant_r}{Quantiles of time to use in constructing the spline, pass in if doing bootstrap.}

\item{nKnots}{Number of knots to use for cubic spline basis (default is 1).}
}
\value{
A list with the elements:
\item{right_dmat}{n*(p+nKnots+2) design matrix for right end of interval.}
\item{left_dmat}{n*(p+nKnots+2) design matrix for left end of interval.}
\item{quant_r}{Quantiles used for constructing spline.}
}
\description{
Puts together the entire design matrix for both the left and right ends of the
interval, pasting together the non-genetic covariates with the cubic spline basis.
}
\examples{
seed <- 2022
n <- 1000
alpha1 <- -0.058
alpha2 <- -0.035
beta1 <- 0.03
beta2 <- log(1 - exp(beta1 / alpha1)) * alpha2
outcomeDat <- genData(seed=seed, n=n, alpha1=alpha1, alpha2=alpha2, beta1=beta1, beta2=beta2)
xMat <- matrix(data=rnorm(n), nrow=n)
obsTimes <- 1:5
lt <- outcomeDat$leftTimes
rt <- outcomeDat$rightTimes
obsInd <- outcomeDat$deltaVecSimple
makeICdmat(xMat = xMat, lt = lt, rt = rt, obsInd = obsInd, quant_r = NULL, nKnots = 1)

}
