% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreEqSpline.R
\name{scoreEqSpline}
\alias{scoreEqSpline}
\title{scoreEqSpline.R}
\usage{
scoreEqSpline(
  x,
  leftDmat,
  rightDmat,
  gSummed = NULL,
  gMat = NULL,
  estG = FALSE,
  leftTimes,
  deltaVec
)
}
\arguments{
\item{x}{(p+nknots+2)*1 vector of fitted coefficients under null model.}

\item{leftDmat}{n*(p+nknots+2) design matrix for left end of interval.}

\item{rightDmat}{n*(p+nknots+2) design matrix for right end of interval.}

\item{gSummed}{n*1 vector of summed genotype matrix (default is NULL).}

\item{gMat}{n*q genotype matrix.}

\item{estG}{Whether to include genotype matrix into design matrix (default is FALSE).}

\item{leftTimes}{n*1 vector of left side of interval times.}

\item{deltaVec}{n*1 vector of the event cause (cause 1 or 2; or right-censored 0).}
}
\value{
A score vector
}
\description{
Calculate score function using cubic spline for baseline cumulative hazard.
}
\examples{
seed <- 2022
n <- 1000
q <- 20
alpha1 <- -0.058
alpha2 <- -0.035
beta1 <- 0.03
beta2 <- log(1 - exp(beta1 / alpha1)) * alpha2
outcomeDat <- genData(seed=seed, n=n, alpha1=alpha1, alpha2=alpha2, beta1=beta1, beta2=beta2)
xMat <- matrix(data=rnorm(n), nrow=n)
gMat <- matrix(data=rbinom(n=n*q, size=2, prob=0.3), ncol=q)
gSummed <- matrix(data=apply(gMat, 1, sum), ncol=1)
obsTimes <- 1:5
lt <- outcomeDat$leftTimes
rt <- outcomeDat$rightTimes
obsInd <- outcomeDat$deltaVecSimple
dmats <- makeICdmat(xMat = xMat, lt = lt, rt = rt, obsInd = obsInd, quant_r = NULL, nKnots = 1)
forUgamma <- scoreEqSpline(x=runif(9, 0.01, 0.02), leftDmat = dmats$left_dmat,
rightDmat = dmats$right_dmat, leftTimes = lt, deltaVec = outcomeDat$deltaVec,
gSummed = gSummed, gMat = gMat, estG = FALSE)

}
